Fix warning in test.

+++ test/unordered_set.cpp
@@ -184,7 +184,7 @@ TEST(uoset_iterate) {
 			}
 			CHECK( count == i );
 
-			s.insert(17 * i);
+			s.insert(int(17 * i));
 		}
 	}
 }
